Move serialization logic to a separate class;
