+++ src/Aer.Memcached.Client/Aer.Memcached.Client.csproj
@@ -30,6 +30,9 @@
             <_Parameter1>Aer.ConsistentHash.Benchmarks</_Parameter1>
         </AssemblyAttribute>
         <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
+            <_Parameter1>Aer.Memcached</_Parameter1>
+        </AssemblyAttribute>
+        <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
             <_Parameter1>Aer.Memcached.Tests</_Parameter1>
         </AssemblyAttribute>
     </ItemGroup>
