+++ src/Aer.Memcached.Client/Interfaces/IObjectBinarySerializer.cs
@@ -1,8 +1,8 @@
 ﻿namespace Aer.Memcached.Client.Interfaces;
 
 /// <summary>
-/// Common interface for binary serializers used to binary serialize
-/// non-primitive typed objects before storing them in memcached.
+/// Common interface for binary serializers used to serialize
+/// objects of non-primitive types before storing them in memcached.
 /// </summary>
 public interface IObjectBinarySerializer
 {
