+++ src/Aer.Memcached.Client/MemcachedClient.cs
@@ -5,9 +5,9 @@ using Aer.ConsistentHash.Abstractions;
 using Aer.Memcached.Client.Commands;
 using Aer.Memcached.Client.Commands.Base;
 using Aer.Memcached.Client.Commands.Enums;
-using Aer.Memcached.Client.Commands.Infrastructure;
 using Aer.Memcached.Client.Interfaces;
 using Aer.Memcached.Client.Models;
+using Aer.Memcached.Client.Serializers;
 using MoreLinq.Extensions;
 
 namespace Aer.Memcached.Client;
@@ -20,20 +20,20 @@ public class MemcachedClient<TNode> : IMemcachedClient where TNode : class, INod
     private readonly ICommandExecutor<TNode> _commandExecutor;
     private readonly IExpirationCalculator _expirationCalculator;
     private readonly ICacheSynchronizer _cacheSynchronizer;
-    private readonly IObjectBinarySerializer _objectBinarySerializer;
+    private readonly BinarySerializer _binarySerializer;
 
     public MemcachedClient(
         INodeLocator<TNode> nodeLocator,
         ICommandExecutor<TNode> commandExecutor,
         IExpirationCalculator expirationCalculator,
         ICacheSynchronizer cacheSynchronizer,
-        IObjectBinarySerializerFactory objectBinarySerializerFactory)
+        BinarySerializer binarySerializer)
     {
         _nodeLocator = nodeLocator;
         _commandExecutor = commandExecutor;
         _expirationCalculator = expirationCalculator;
         _cacheSynchronizer = cacheSynchronizer;
-        _objectBinarySerializer = objectBinarySerializerFactory.Create();
+        _binarySerializer = binarySerializer;
     }
 
     /// <inheritdoc />
@@ -52,7 +52,7 @@ public class MemcachedClient<TNode> : IMemcachedClient where TNode : class, INod
                 return MemcachedClientResult.Unsuccessful($"Memcached node for key {key} is not found");
             }
 
-            var cacheItem = BinaryConverter.Serialize(value, _objectBinarySerializer);
+            var cacheItem = _binarySerializer.Serialize(value);
             var expiration = _expirationCalculator.GetExpiration(key, expirationTime);
 
             using (var command = new StoreCommand(storeMode, key, cacheItem, expiration))
@@ -196,8 +196,7 @@ public class MemcachedClient<TNode> : IMemcachedClient where TNode : class, INod
                 }
 
                 var deserializationResult =
-                    BinaryConverter.Deserialize<T>(commandExecutionResult.GetCommandAs<GetCommand>().Result,
-                        _objectBinarySerializer);
+                    _binarySerializer.Deserialize<T>(commandExecutionResult.GetCommandAs<GetCommand>().Result);
 
                 return MemcachedClientValueResult<T>.Successful(
                     deserializationResult.Result,
@@ -289,7 +288,7 @@ public class MemcachedClient<TNode> : IMemcachedClient where TNode : class, INod
                 var key = readItem.Key;
                 var cacheItem = readItem.Value;
 
-                var cachedValue = BinaryConverter.Deserialize<T>(cacheItem, _objectBinarySerializer).Result;
+                var cachedValue = _binarySerializer.Deserialize<T>(cacheItem).Result;
 
                 result.TryAdd(key, cachedValue);
             }
@@ -530,7 +529,7 @@ public class MemcachedClient<TNode> : IMemcachedClient where TNode : class, INod
 
                 foreach (var key in keys)
                 {
-                    keyValuesToStore[key] = BinaryConverter.Serialize(keyValues[key], _objectBinarySerializer);
+                    keyValuesToStore[key] = _binarySerializer.Serialize(keyValues[key]);
                 }
 
                 var command = new MultiStoreCommand(storeMode, keyValuesToStore, keyToExpirationMap);
@@ -584,7 +583,7 @@ public class MemcachedClient<TNode> : IMemcachedClient where TNode : class, INod
                         var keyValuesToStore = new Dictionary<string, CacheItemForRequest>();
                         foreach (var key in keysBatch)
                         {
-                            keyValuesToStore[key] = BinaryConverter.Serialize(keyValues[key], _objectBinarySerializer);
+                            keyValuesToStore[key] = _binarySerializer.Serialize(keyValues[key]);
                         }
 
                         using (var command = new MultiStoreCommand(storeMode, keyValuesToStore, keyToExpirationMap))
@@ -643,7 +642,7 @@ public class MemcachedClient<TNode> : IMemcachedClient where TNode : class, INod
                         var key = readItem.Key;
                         var cacheItem = readItem.Value;
 
-                        var cachedValue = BinaryConverter.Deserialize<T>(cacheItem, _objectBinarySerializer).Result;
+                        var cachedValue = _binarySerializer.Deserialize<T>(cacheItem).Result;
 
                         result.TryAdd(key, cachedValue);
                     }
