+++ src/Aer.Memcached.Client/Serializers/Binary/BsonObjectBinarySerializer.cs
