+++ src/Aer.Memcached.Client/Serializers/Json/DefaultJsonSerializer.cs
@@ -1,4 +1,4 @@
-﻿using Aer.Memcached.Client.Commands.Infrastructure.JsonConverters;
+﻿using Aer.Memcached.Client.Serializers.JsonConverters;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Serialization;
