+++ tests/Aer.Memcached.Tests/TestClasses/MemcachedMaintainerTests.cs
@@ -8,7 +8,6 @@ using Aer.Memcached.Diagnostics.Listeners;
 using Aer.Memcached.Infrastructure;
 using Aer.Memcached.Tests.Infrastructure;
 using Aer.Memcached.Tests.Infrastructure.Logging;
-using Aer.Memcached.Tests.Model;
 using FluentAssertions;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
